test firts part implemented. minor class refactoring improvements
