+++ src/main/java/com/coffeecart/ui/page/MenuPage.java
@@ -3,15 +3,12 @@ package com.coffeecart.ui.page;
 import com.coffeecart.ui.component.CardComponent;
 import com.coffeecart.ui.component.LuckyDayComponent;
 import com.coffeecart.ui.elements.TotalButtonElement;
-import com.coffeecart.ui.elements.TotalButtonElement;
-import com.coffeecart.ui.modal.LuckyDayModal;
 import com.coffeecart.ui.modal.PaymentDetailModal;
 import io.qameta.allure.Step;
 
 import lombok.Getter;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 
 import java.util.ArrayList;
@@ -23,20 +20,16 @@ public class MenuPage extends BasePage {
     private List<CardComponent> cards = new ArrayList<>();
 
     @Getter
-    @FindBy(xpath = "//*[@id='app']/div[2]/ul/li")
-    private List<WebElement> rootCards;
+    @FindBy(xpath="//*[@id='app']/div[2]/ul/li")
+    private List<WebElement> rootCards ;
 
     @Getter
-    @FindBy(xpath = "//div[@class='modal']/div[@class='modal-content size']")
-    private WebElement paymentModalRoot;
+    @FindBy(xpath="//div[@class='modal']/div[@class='modal-content size']")
+    private WebElement paymentModalRoot ;
 
     @Getter
-    @FindBy(xpath = "//div[@class='promo']")
-    private WebElement luckyDayModalRoot;
-
-    @Getter
-    @FindBy(xpath = "//*[@class=\"pay-container\"]")
-    private WebElement payContainer;
+    @FindBy(xpath="//div[@class='promo']")
+    private WebElement luckyDayModalRoot ;
 
     @Getter
     @FindBy(xpath="//button[@class='pay']")
@@ -45,16 +38,10 @@ public class MenuPage extends BasePage {
     @Getter
     @FindBy(xpath="//*[@class=\"pay-container\"]")
     private WebElement payContainer ;
-    @FindBy(xpath = "//button[@class='pay']")
-    private WebElement totalButton;
 
     public MenuPage(WebDriver driver) {
         super(driver);
-        totalButton = new TotalButtonElement(driver,totalButtonRoot);
         for(WebElement card: rootCards) {
-        luckyDayModal = new LuckyDayModal(driver, luckyDayModalRoot);
-        paymentDetailModal = new PaymentDetailModal(driver, paymentModalRoot);
-        for (WebElement card : rootCards) {
             cards.add(new CardComponent(driver, card));
         }
     }
@@ -65,7 +52,7 @@ public class MenuPage extends BasePage {
     }
 
     public TotalButtonElement getButtonElement(){
-        return totalButton;
+        return new TotalButtonElement(driver,totalButtonRoot);
     }
 
     public LuckyDayComponent getGetLackyDayComponent(){
@@ -89,18 +76,6 @@ public class MenuPage extends BasePage {
     @Step("Navigate to the GitHub Page")
     public GitHubPage goToGitHubPage() {
         return header.navigateToGitHub();
-    }
-
-    public TotalButtonElement getButtonElement() {
-        return new TotalButtonElement(driver, getPayContainer());
-    }
-
-    public MenuPage clickDrink(String drinkName) {
-        getCards().stream()
-                .filter(card -> card.getName().equals(drinkName))
-                .findFirst()
-                .ifPresent(component -> component.getCupComponent().getCupBody().click());
-        return this;
     }
 }
